+++ cdm/core/src/main/java/ucar/nc2/dataset/VariableDS.java
@@ -8,9 +8,11 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Sets;
 import ucar.ma2.*;
 import ucar.nc2.*;
+import ucar.nc2.constants.AxisType;
 import ucar.nc2.constants.CDM;
 import ucar.nc2.dataset.NetcdfDataset.Enhance;
 import ucar.nc2.internal.dataset.CoordinatesHelper;
+import ucar.nc2.time.CalendarDate;
 import ucar.nc2.util.CancelTask;
 import javax.annotation.Nullable;
 import java.io.IOException;
