+++ cdm/core/src/test/data/ncml/aggSynGrid.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <netcdf xmlns="http://www.unidata.ucar.edu/namespaces/netcdf/ncml-2.2">
 
-  <aggregation  dimName="time" type="joinNew">
+  <aggregation dimName="time" type="joinNew">
     <variableAgg name="T"/>
     <netcdf location="nc/time0.nc" coordValue="2005-11-22 22:19:53Z"/>
     <netcdf location="nc/time1.nc" coordValue="2005-11-22 23:19:53Z"/>
