--- cdm/core/src/test/data/testUnsignedFillValue.dump
@@ -1,13 +1,0 @@
-netcdf file:../core/src/test/data/testUnsignedFillValue.ncml {
-  variables:
-    uint my_var;
-      :long_name = "my_var_name";
-      :_FillValue = 55555US; // ushort
-      :_Unsigned = "true";
-
-  // global attributes:
-  :_CoordSysBuilder = "ucar.nc2.dataset.conv.DefaultConvention";
-
-  data:
-    my_var = 55555
-}
