+++ cdm/core/src/test/java/ucar/nc2/dataset/conv/TestDefaultCalendars.java
@@ -24,8 +24,8 @@ public class TestDefaultCalendars {
   private final Calendar defaultCoardsCalendar = Calendar.gregorian;
   private final Calendar defaultCFCalendar = Calendar.gregorian;
 
-  private final String coardsConvention = COARDSConvention.class.getName();
-  private final String cfConvention = CF1Convention.class.getName();
+  private final String coardsConvention = "COARDS";
+  private final String cfConvention = "CF-1.X";
 
 
   @Test
