+++ cdm/core/src/test/java/ucar/nc2/dataset/conv/TestSimpleGeom.java
@@ -16,7 +16,7 @@ import static java.lang.String.format;
 
 public class TestSimpleGeom {
 
-  private final String cfConvention = CF1Convention.class.getName();
+  private final String cfConvention = "CF-1.X";
 
   @Test
   public void testLine() throws IOException {
